🎨 move common code into tracker. update cli.py
